- 3D 7-point and 27-point stencil CG solvers
- Compute-communication overlap with dual-stream execution
- Z-slab partitioning for 3D grids
- Benchmarks on 8x A100 (88% scaling efficiency, up to 1.45x overlap gain)
- Nsight Systems timeline screenshots in README
- 3D scaling and overlap gain plots

+++ include/spmv_csr.h
@@ -28,8 +28,8 @@
 struct CSRMatrix {
     int nb_rows;
     int nb_cols;
-    int nb_nonzeros;
-    int* row_ptr;
+    long long nb_nonzeros;
+    long long* row_ptr;
     int* col_indices;
     double* values;
 };
